Moved document type to types file
